Worklog.Attachment

Description

The Attachment type describes a document related to the case. The document's minimal

Documentation

data Attachment

Very simple type, Urgency and a Summary type to hold the todos matter.

Constructors

Attachment

Fields

attachmentKind :: AttachmentKind

attachmentPath :: Maybe RelativePath

attachmentIds :: Maybe [DocumentID]

attachmentMimeType :: Maybe MimeType

attachmentSubject :: DocumentSubjet

data AttachmentKind

A Todo can be Urgent, which means that it must be executed as soon as possible, today, or tomorrow. NotSoUrgent Todos can postpone for a while.

Constructors

Picture

Incoming IncomingDocType

Outgoing OutgoingDocType

Internal InternalDocType

GeneralDoc

data IncomingDocType

data OutgoingDocType

data InternalDocType

newAttachment :: AttachmentKind -> Maybe RelativePath -> Maybe [DocumentID] -> DocumentSubjet -> Attachment