Worklog.Case

Description

Domain logic for the individual cases.

Task and its attributes

data Case

The Case data type contains the parameters of the case correspondent to the workfile. Some title, a longer summary, priority and status, and of course deadline. The caseFlags attributes

The summary field currently will be implemented as Pandoc Block.

The case has an ID.

caseId :: Case -> Int

caseTitle :: Case -> String

casePriority :: Case -> Priority

caseDeadline :: Case -> Deadline

newCase :: Int -> String -> Case

mkTask id title creates an open task with normal priority, no deadline and empty summary.

Status of task

data Status

Constructors

OnDesk
case is on the desk, we work on it

Deferred Day
case is in the cabinet, waiting for the deadline to come back on desk

Archived
case is closed, it is in the archives, hoppefully never comes back

isOnDesk :: Case -> Bool

isArchived :: Case -> Bool

Checks for finished case

isDeferred :: Case -> Bool

getDefferedDay :: Case -> Maybe Day

deferrCase :: Case -> Day -> Case

Priority

data Priority

Constructors

ForgetIt

Postpone

Normal

Important

VeryImportant

isImportant :: Case -> Bool

needsAttention :: Case -> Bool

Not closed and high priority case

setCasePriority :: Priority -> Case -> Case

promoteCase :: Case -> Case

demoteCase :: Case -> Case

Deadlines

data Deadline

the deadline of the task

Constructors

NoDeadline
For tasks without deadline

Deadline Day
Deadline

hasDeadline :: Case -> Bool

setDeadline :: Deadline -> Case -> Case

Flags

data CaseFlag

Flags are attributes of the task.

Constructors

Waiting
The task is waiting for someone, something. Work cannot continue, until it has not arrived.

NeedsManager
The task needs decision from the boss.

ApprovalPending

anyFlag :: Case -> Bool

hasFlag :: CaseFlag -> Case -> Bool

addFlag :: CaseFlag -> Case -> Case

removeFlag :: CaseFlag -> Case -> Case

Other types

type Warning = String

Functions

closeCase :: Case -> Either Warning Case

openCase :: Case -> Case