Worklog.Implementation.Internal

Description

The internal implementations This modul exports the types and functions to handle the hidden implementations.

Summary type

data Summary

The Summary type is used by more types in the application. It is a placeholder of a longer text attribute of the types. Its internal implementation is hidden from the types. Generally the parser and the renderer types need the functions to handle the Summary values.

Currently the internal representation is String. Todo: Use Text or Pandoc [Block] Todo: This module can contain some usefull functions to work with Summary in a modul which cannot use the internal representation.

mkSummary :: String -> Summary

summaryText :: Summary -> String

emptySummary :: Summary

mapSummary :: (String -> String) -> Summary -> Summary

HasSummary type class

class HasSummary a where

The HasSummary is the type class of the types which want to use the Summary type. Because the Summary must be hidden to them, there must be the class methods to read, set and modify the parameter.

Minimal complete definition

setSummary, getSummary

Methods

getSummary :: a -> Summary

setSummary :: a -> Summary -> a

modifySummary :: a -> (Summary -> Summary) -> a

appendSummary :: a -> Summary -> a