Worklog.Task

Description

Task logic

Task and its attributes

data Task

The Task data type contains the parameters of the case correspondent to the workfile. Some title, a longer summary, priority and status, and of course deadline.

The summary field currently implemented as Pandoc Block.

The task has an ID.

taskId :: Task -> Int

taskTitle :: Task -> String

taskPriority :: Task -> Priority

taskStatus :: Task -> Status

taskDeadline :: Task -> Deadline

mkTask :: Int -> String -> Task

mkTask id title creates an open task with normal priority, no deadline and empty summary.

Status of task

data Status

Constructors

Open
task is under work

Reviewed
task is reviewed by others, it is ready to make the decision

Closed
task is closed, there is no work with it

isOpen :: Task -> Bool

isFinished :: Task -> Bool

Checks for finished task

Priority

data Priority

Constructors

ForgetIt

Postpone

Normal

Important

VeryImportant

isImportant :: Task -> Bool

needsAttention :: Task -> Bool

Not closed and high priority task

setTaskPriority :: Priority -> Task -> Task

promoteTask :: Task -> Task

demoteTask :: Task -> Task

Deadlines

data Deadline

Very important attribute.

Constructors

NoDeadline
For tasks without deadline

Deadline Day
Deadline

hasDeadline :: Task -> Bool

setDeadline :: Deadline -> Task -> Task

Functions

closeTask :: Task -> Task

openTask :: Task -> Task

mkSummary :: String -> Summary

summaryText :: Summary -> String

emptySummary :: Summary

getSummary :: Task -> Summary

setSummary :: Summary -> Task -> Task